Use non-UInt32 in attachment en-/decryption

+++ SignalServiceKit/Backups/Archiving/BackupArchiveAttachmentByteCounter.swift
@@ -9,9 +9,9 @@ public class BackupArchiveAttachmentByteCounter {
     private var bytesCounter: UInt64 = 0
     private var includedAttachmentsInByteCount: Set<Attachment.IDType> = Set()
 
-    func addToByteCount(attachmentID: Attachment.IDType, byteCount: UInt32) {
+    func addToByteCount(attachmentID: Attachment.IDType, byteCount: UInt64) {
         if includedAttachmentsInByteCount.insert(attachmentID).inserted {
-            bytesCounter += UInt64(byteCount)
+            bytesCounter += byteCount
         }
     }
